refactor(inference): chnages folder name,  delete additional prints, add model without initialization
#11
